DA#229: Modified: Added license and created directory types
