+++ src/extension.ts
@@ -19,9 +19,9 @@ import { ClusterConnectionNode } from "./model/ClusterConnectionNode";
 import CollectionNode from "./model/CollectionNode";
 import DocumentNode from "./model/DocumentNode";
 import { DocumentNotFoundError } from "couchbase";
-import { IConnection } from "./model/IConnection";
-import { INode } from "./model/INode";
-import { logger } from "./logging/logger";
+import { IConnection } from "./types/IConnection";
+import { INode } from "./types/INode";
+import { logger } from "./logger/logger";
 import { PagerNode } from "./model/PagerNode";
 import { ScopeNode } from "./model/ScopeNode";
 import { getBucketMetaData, getDocumentMetaData } from "./webViews/metaData.webview";
@@ -35,7 +35,7 @@ import {
 } from "./util/connections";
 import { MemFS } from "./util/fileSystemProvider";
 import { Global, Memory, WorkSpace } from "./util/util";
-import { IDocumentData } from "./model/IDocument";
+import { IDocumentData } from "./types/IDocument";
 import { QueryContentSerializer } from "./notebook/serializer";
 import { QueryKernel } from "./notebook/controller";
 import { Constants } from "./util/constants";
