+++ src/logger/enum.ts
@@ -13,29 +13,10 @@
  *   See the License for the specific language governing permissions and
  *   limitations under the License.
  */
-import * as vscode from "vscode";
-import CollectionNode from "./CollectionNode";
-import { INode } from "../types/INode";
 
-export class PagerNode implements INode {
-
-  constructor(
-    public readonly collection: CollectionNode
-  ) { }
-
-  public getTreeItem(): vscode.TreeItem {
-    return {
-      label: `Load more`,
-      collapsibleState: vscode.TreeItemCollapsibleState.None,
-      command: {
-        command: "vscode-couchbase.loadMore",
-        title: "Load more",
-        arguments: [this],
-      },
-    };
-  }
-
-  public async getChildren(): Promise<INode[]> {
-    return [];
-  }
+export enum Level {
+    debug = "DEBUG",
+    info = "INFO",
+    warn = "WARN",
+    error = "ERROR"
 }
