--- src/logging/enum.ts
@@ -1,6 +1,0 @@
-export enum Level {
-    debug = "DEBUG",
-    info = "INFO",
-    warn = "WARN",
-    error = "ERROR"
-}
