+++ src/types/IConnection.ts
