+++ src/types/IDocument.ts
