+++ src/types/INode.ts
